ci: prevent redundant workflow runs.
Lint and test workflows were running two times.
(1) When a pull request was submitted, to verify that it was okay
to merge into main.
(2) After a pull request was merged, as a check before building
and releasing the next version of the package on PyPi.
The second run is unnecessary, as (1) ensures that (2) will always
pass. This commit removes (2).