feat(branch_and_bound): add n-best functionality.
`pykp.solvers.branch_and_bound` now accepts an optional argument
`n`, which specifies the number of 'best' solutions to return.
The `n` argument is on solution values, not the number of
solutions. If `n` is set to 1, the solver returns all solutions
that achieve the distinct optimal value. More than one solution
may be returned if there are multiple solutions with the same
optimal value. Similarly, if `n` is set to n, the solver returns
all solutions that achieve the n-highest possible values.