fix(phase_transition): dimensions back to front.
The final `phase_transition` array was being reshaped in a manner
that was inconsistent with the way in which  resolution was passed
to the function. The expected outcome should be to reshape the
phase transition in such a way that nc varies along the x-axis
(column-wise) and np along the y-axis (row-wise).