feat(sampler): make reproducible.
`pykp.sampler.Sampler.sample()` now accepts a `seed` argument which
ensures samples are reproducible. Stop storing already-bound
generator methods in the constructor. Instead, store the
distribution name and kwargs, which are initialised in a seperate
private method. This allows a new `rng` with a provided `seed`
to bne generated in teh `sample` method.